+++ src/Application/RoadGeocoderInterface.php
@@ -30,10 +30,6 @@ interface RoadGeocoderInterface
 
     public function findSides(string $administrator, string $roadNumber, ?string $departmentCode, string $pointNumber): array;
 
-    public function findNamedStreets(string $search, string $cityCode): array;
-
-    public function findRoadNames(string $search, string $cityCode): array;
-
     public function findSectionsInArea(string $areaGeometry, array $excludeTypes = [], ?bool $clipToArea = false): string;
 
     public function convertPolygonRoadToLines(string $geometry): string;
