+++ src/Infrastructure/Controller/Regulation/Fragments/GetAddressCompletionFragmentController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Controller\Regulation\Fragments;
 
-use App\Application\RoadGeocoderInterface;
+use App\Application\GeocoderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
@@ -13,7 +13,7 @@ use Symfony\Component\Routing\Annotation\Route;
 final class GetAddressCompletionFragmentController
 {
     public function __construct(
-        private RoadGeocoderInterface $roadGeocoder,
+        private GeocoderInterface $geocoder,
         private \Twig\Environment $twig,
     ) {
     }
@@ -32,7 +32,7 @@ final class GetAddressCompletionFragmentController
             throw new BadRequestHttpException();
         }
 
-        $namedStreets = $this->roadGeocoder->findNamedStreets($search, $cityCode);
+        $namedStreets = $this->geocoder->findNamedStreets($search, $cityCode);
 
         return new Response(
             $this->twig->render(
